Initial LLVM integration work.
Not all instructions fleshed out in LLVM yet. Expect breakages.

+++ vm/builtin_contexts.hpp
@@ -24,7 +24,7 @@ namespace rubinius {
 
     Tuple* stack; // slot
 
-    size_t ip;
+    int    ip;
     int    sp;
     size_t args;
     OBJECT block; // slot
